+++ composeApp/src/commonMain/kotlin/com/poli/health/aquamate/intake/domain/usecase/DeleteIntakeByIdUseCase.kt
@@ -1,0 +1,7 @@
+package com.poli.health.aquamate.intake.domain.usecase
+
+import kotlinx.datetime.LocalDate
+
+interface DeleteIntakeByIdUseCase {
+    suspend operator fun invoke(userId: String, date: LocalDate, intakeId: String): Result<Unit>
+}